requirements.txt
flask==2.3.3
gtts==2.3.2
python-dotenv==1.0.0
requests==2.31.0
SpeechRecognition==3.10.0
pyaudio==0.2.13
numpy>=1.21.0
sounddevice>=0.4.5
soundfile>=0.10.3
pydub>=0.25.1
python-socketio>=5.7.2
python-engineio>=4.4.1
python-rtmidi>=1.4.9
webrtcvad>=2.0.10
openai-whisper>=20230314
edge-tts>=6.1.3
